removed old api code and combined old server with new server
