+++ lib/api.js
@@ -12,6 +12,17 @@ function isFunction (f) {
   return 'function' === typeof f
 }
 
+function toBuffer(v) {
+  if (Buffer.isBuffer(v)) return v
+  if (typeof v == 'string') return new Buffer(v, 'hex')
+  return new Buffer(v)
+}
+
+function toHexString(b) {
+  if (Buffer.isBuffer(b)) return b.toString('hex')
+  return b.toString()
+}
+
 var manifest = {
   async: [
     'add',
@@ -19,7 +30,13 @@ var manifest = {
     'getPublicKey',
     'getLatest',
     'whoami',
-    'auth'
+    'auth',
+
+    // admin api
+    'follow',
+    'unfollow',
+    'isFollowing',
+    'setProfile'
   ],
 
   source: [
@@ -31,7 +48,10 @@ var manifest = {
     'messagesLinkedToFeed',
     'messagesLinkedFromFeed',
     'feedsLinkedToFeed',
-    'feedsLinkedFromFeed'
+    'feedsLinkedFromFeed',
+
+    // admin api
+    'followedUsers'
   ]
 }
 
@@ -75,6 +95,56 @@ function api (ssb, feed, config) {
       return cb()
     }
     cb(new Error('access denied'))
+  }
+
+  api.setProfile = function(prof, cb) {
+    if (!prof || typeof prof != 'object' || !prof.nickname || typeof prof.nickname != 'string')
+      return cb(new Error('Must pass an object with {nickname: String}'))
+    feed.add({ type: 'profile', nickname: prof.nickname }, cb)
+  }
+
+  api.follow = function(id, cb) {
+    id = toBuffer(id)
+    if (!id) return cb(new Error('Invalid ID'))
+
+    // publish follows link
+    feed.add({ type: 'follow', $feed: id, $rel: 'follows' }, cb)
+  }
+
+  api.unfollow = function(id, cb) {
+    id = toBuffer(id)
+    if (!id) return cb(new Error('Invalid ID'))
+
+    api.isFollowing(id, function(err, isFollowing) {
+      if (err) return cb(err)
+      if (!isFollowing) return cb()
+
+      // publish unfollows link
+      feed.add({ type: 'follow', $feed: id, $rel: 'unfollows' }, cb)
+    })
+  }
+
+  api.isFollowing = function(id, cb) {
+    var id = toHexString(id)
+    pull(api.followedUsers(), pull.drain(
+      function(id2) {
+        if (id == toHexString(id2)) {
+          cb(null, true)
+          return false // done processing stream
+        }
+      },
+      function(aborted) {
+        if (!aborted)
+          cb(null, false)
+      }
+    ))
+  }
+
+  api.followedUsers = function() {
+    return pull(
+      ssb.feedsLinkedFromFeed(feed.id, 'follows'),
+      pull.map(function(link) { return link.dest })
+    )
   }
 
   return api
